requirements.txt
openai>=1.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0
streamlit>=1.35.0
plotly>=5.20.0
python-dotenv>=1.0.0
httpx>=0.27.0
